Add asynchronours requests to the lobby state machine
Instead of getting new data from the server every second, we now fetch
new data asynchronously using http long polling on the server. To
implement this, a counter is added to the lobby function that
effectively acts as a mutex. This counter is unset whenever the function
returns.

If nothing gets returned by the server other than the 503 timeout
response, the function is called again recursively until it returns with
new data. During these recursive calls the funtion remains blocked to
the timer triggered function calls.